backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
dotenv==0.9.9
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.5.1